Change fetch stream to continue
The kraken fetch stream stops when an error is returned in the
reponse message itself (everything else was successful). This could
be the server temporarily going down or a miconfiguration error or
something else. Either way, things might work again or human
intervention is required. Default to trusting that things might
work again.