Get last timestamp on startup
When starting up, will fetch the last records for each asset pair
and set the timestamp. This will ensure that the first PUT request
won't cause existing trade history items to be doubled up in the
database. This would be a huge problem if collector is restarted
frequently. The kraken fetcher fetches the last 1000 trade history
items even if there were one or two new items or none each time.